1.0b Update
Instead of subscribing to debug logs, the console is explicitly called by the UI manager to update its text.
Ive also added a bool to the UI manager to disallow interacting with the lock when the credits popup is active.
Again, only the LockPicking_Standalone_Demo scene is considered stable.